fix(sync): upload profile changes automatically
